@@ -1,6 +1,7 @@
 package com.xingray.csv;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.Date;
